refactor(support-plan-guide): split presentation helpers from guide page
